Allow matching multiple message types
Often the type in `.hears` can be multiple things. Botkit supports both arrays of types and comma-separated strings so I've added in basic support for both those situations so the `.hears` still gets triggered when the two types aren't exactly equal. 

Casting to an array as opposed to a string so indexOf will match the whole word only (stops mention and direct_mention being confused). Not sure if there's a more efficient way to do this with regex.